Release 0.1.2
