+++ core/rinux/Cargo.toml
@@ -23,22 +23,55 @@
 # 
 
 [package]
-name = "rinux"
+name = "rinuxcore"
 version = "0.1.2"
-authors = ["AtomicGamer@github.com"]
+authors = ["AtomicGamer9523@github.com"]
 edition = "2018"
 license = "MIT"
 
+[[test]]
+name = "should_panic"
+harness = false
 
+[[test]]
+name = "stack_overflow"
+harness = false
 
+[features]
+epearl = []
 
-[build-dependencies.epearl]
-path = "./core/enderpearl"
-package = "enderpearl"
+[dependencies]
+bootloader = { version = "0.9.8", features = ["map_physical_memory"]}
+epearl = { path = "../enderpearl", package = "enderpearl" }
+volatile = "0.2.6"
+spin = "0.5.2"
+x86_64 = "0.14.2"
+uart_16550 = "0.2.0"
+pic8259 = "0.10.1"
+pc-keyboard = "0.5.0"
+linked_list_allocator = "0.9.0"
 
-[dependencies.rinuxcore]
-path = "./core/rinux"
-package = "rinuxcore"
-features = [
-    "epearl"
+[dependencies.lazy_static]
+version = "1.0"
+features = ["spin_no_std"]
+
+[dependencies.crossbeam-queue]
+version = "0.2.1"
+default-features = false
+features = ["alloc"]
+
+[dependencies.conquer-once]
+version = "0.2.0"
+default-features = false
+
+[dependencies.futures-util]
+version = "0.3.4"
+default-features = false
+features = ["alloc"]
+
+[package.metadata.bootimage]
+test-args = [
+    "-device", "isa-debug-exit,iobase=0xf4,iosize=0x04", "-serial", "stdio",
+    "-display", "none"
 ]
+test-success-exit-code = 33
