+++ core/rinux/src/conf/file.rs
@@ -1,0 +1,4 @@
+#![allow(dead_code)]
+pub(crate) const PROJECT_NAME: &'static str = "MyProject";
+pub(crate) const PROJECT_VERSION: &'static str = "v0.1.0";
+pub(crate) const QUIET_BOOT: bool = false;
