--- rinux/enderpearl/Cargo.toml
@@ -1,5 +1,0 @@
-[package]
-name = "enderpearl"
-version = "0.1.1"
-edition = "2021"
-authors = ["AtomicGamer9523@github.com"]
