--- rinux/enderpearl/src/runner.rs
@@ -1,5 +1,0 @@
-extern crate alloc;use alloc::string::String;
-
-pub(crate) fn run(_command: &String, _print: bool) {
-    todo!();
-}
