--- rinux/src/conf/mod.rs
@@ -1,4 +1,0 @@
-#![allow(dead_code)]
-    pub(crate) const PROJECT_NAME: &'static str = "MyProject";
-    pub(crate) const PROJECT_VERSION: &'static str = "v0.1.0";
-    pub(crate) const QUIET_BOOT: bool = false;
