--- rinux/src/vga_buffer.rs
@@ -1,323 +1,0 @@
-use core::fmt;
-use lazy_static::lazy_static;
-use spin::Mutex;
-use volatile::Volatile;
-
-lazy_static! {
-    pub(crate) static ref WRITER: Mutex<Writer> = Mutex::new(Writer {
-        column_position: 0,
-        color_code: ColorCode::new(Color::White, Color::Black),
-        buffer: unsafe { &mut *(0xb8000 as *mut Buffer) },
-    });
-
-    pub(crate) static ref INFOWRITER: Mutex<Writer> = Mutex::new(Writer {
-        column_position: 0,
-        color_code: ColorCode::new(Color::Magenta, Color::Black),
-        buffer: unsafe { &mut *(0xb8000 as *mut Buffer) },
-    });
-
-    pub(crate) static ref LOGOWRITER: Mutex<Writer> = Mutex::new(Writer {
-        column_position: 0,
-        color_code: ColorCode::new(Color::Brown, Color::Black),
-        buffer: unsafe { &mut *(0xb8000 as *mut Buffer) },
-    });
-
-    pub(crate) static ref OKWRITER: Mutex<Writer> = Mutex::new(Writer {
-        column_position: 0,
-        color_code: ColorCode::new(Color::Green, Color::Black),
-        buffer: unsafe { &mut *(0xb8000 as *mut Buffer) },
-    });
-
-    pub(crate) static ref WARNWRITER: Mutex<Writer> = Mutex::new(Writer {
-        column_position: 0,
-        color_code: ColorCode::new(Color::LightRed, Color::Black),
-        buffer: unsafe { &mut *(0xb8000 as *mut Buffer) },
-    });
-
-    pub(crate) static ref ERRWRITER: Mutex<Writer> = Mutex::new(Writer {
-        column_position: 0,
-        color_code: ColorCode::new(Color::Red, Color::Black),
-        buffer: unsafe { &mut *(0xb8000 as *mut Buffer) },
-    });
-}
-
-
-#[allow(dead_code)]
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-#[repr(u8)]
-pub enum Color {
-    Black = 0,
-    Blue = 1,
-    Green = 2,
-    Cyan = 3,
-    Red = 4,
-    Magenta = 5,
-    Brown = 6,
-    LightGray = 7,
-    DarkGray = 8,
-    LightBlue = 9,
-    LightGreen = 10,
-    LightCyan = 11,
-    LightRed = 12,
-    Pink = 13,
-    Yellow = 14,
-    White = 15,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-#[repr(transparent)]
-struct ColorCode(u8);
-
-impl ColorCode {
-    fn new(foreground: Color, background: Color) -> ColorCode {
-        ColorCode((background as u8) << 4 | (foreground as u8))
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-#[repr(C)]
-struct ScreenChar {
-    ascii_character: u8,
-    color_code: ColorCode,
-}
-
-const BUFFER_HEIGHT: usize = 25;
-const BUFFER_WIDTH: usize = 80;
-
-#[repr(transparent)]
-struct Buffer {
-    chars: [[Volatile<ScreenChar>; BUFFER_WIDTH]; BUFFER_HEIGHT],
-}
-
-pub(crate) struct Writer {
-    column_position: usize,
-    color_code: ColorCode,
-    buffer: &'static mut Buffer,
-}
-impl Writer {
-    pub(crate) fn write_byte(&mut self, byte: u8) {
-        match byte {
-            b'\n' => self.new_line(),
-            byte => {
-                if self.column_position >= BUFFER_WIDTH {
-                    self.new_line();
-                }
-                let row = BUFFER_HEIGHT - 1;
-                let col = self.column_position;
-                let color_code = self.color_code;
-                self.buffer.chars[row][col].write(ScreenChar {
-                    ascii_character: byte,
-                    color_code,
-                });
-                self.column_position += 1;
-            }
-        }
-    }
-    fn write_string(&mut self, s: &str) {
-        for byte in s.bytes() {
-            match byte {
-                0x20..=0x7e | b'\n' => self.write_byte(byte),
-                _ => self.write_byte(0xfe),
-            }
-        }
-    }
-    fn new_line(&mut self) {
-        for row in 1..BUFFER_HEIGHT {
-            for col in 0..BUFFER_WIDTH {
-                let character = self.buffer.chars[row][col].read();
-                self.buffer.chars[row - 1][col].write(character);
-            }
-        }
-        self.clear_row(BUFFER_HEIGHT - 1);
-        self.column_position = 0;
-    }
-    fn clear_row(&mut self, row: usize) {
-        let blank = ScreenChar {
-            ascii_character: b' ',
-            color_code: ColorCode::new(Color::Black,Color::Black),
-        };
-        for col in 0..BUFFER_WIDTH {
-            self.buffer.chars[row][col].write(blank);
-        }
-    }
-}
-impl fmt::Write for Writer {
-    fn write_str(&mut self, s: &str) -> fmt::Result {
-        self.write_string(s);
-        Ok(())
-    }
-}
-
-#[macro_export]
-macro_rules! println {
-    (
-
-    ) => (
-        $crate::print!(
-            "\n"
-        )
-    );
-
-    (
-        $(
-            $arg: tt
-        )*
-    ) => (
-        $crate::print!(
-            "{}\n",
-            format_args!(
-                $(
-                    $arg
-                )*
-            )
-        )
-    );
-}
-
-#[macro_export]
-macro_rules! print {
-    (
-        $(
-            $arg: tt
-        )*
-    ) => (
-        $crate::vga_buffer::_print(
-            format_args!(
-                $(
-                    $arg
-                )*
-            )
-        )
-    );
-}
-
-#[macro_export]
-macro_rules! print_ok {
-    (
-        $(
-            $arg: tt
-        )*
-    ) => (
-        $crate::vga_buffer::_print_ok(
-            format_args!(
-                $(
-                    $arg
-                )*
-            )
-        )
-    );
-}
-
-#[macro_export]
-macro_rules! print_err {
-    (
-        $(
-            $arg: tt
-        )*
-    ) => (
-        $crate::vga_buffer::_print_err(
-            format_args!(
-                $(
-                    $arg
-                )*
-            )
-        )
-    );
-}
-
-#[macro_export]
-macro_rules! print_info {
-    (
-        $(
-            $arg: tt
-        )*
-    ) => (
-        $crate::vga_buffer::_print_info(
-            format_args!(
-                $(
-                    $arg
-                )*
-            )
-        )
-    );
-}
-
-
-#[doc(hidden)]
-pub fn _print(args: fmt::Arguments) {
-    use core::fmt::Write;
-    use x86_64::instructions::interrupts;
-    interrupts::without_interrupts(|| {
-        WRITER.lock().write_fmt(args).unwrap();
-    });
-}
-
-#[doc(hidden)]
-pub(crate) fn _print_logo(args: fmt::Arguments) {
-    use core::fmt::Write;
-    use x86_64::instructions::interrupts;
-    interrupts::without_interrupts(|| {
-        LOGOWRITER.lock().write_fmt(args).unwrap();
-    });
-}
-
-#[doc(hidden)]
-pub fn _print_ok(args: fmt::Arguments) {
-    use core::fmt::Write;
-    use x86_64::instructions::interrupts;
-    interrupts::without_interrupts(|| {
-        OKWRITER.lock().write_fmt(args).unwrap();
-    });
-}
-
-#[doc(hidden)]
-pub fn _print_err(args: fmt::Arguments) {
-    use core::fmt::Write;
-    use x86_64::instructions::interrupts;
-    interrupts::without_interrupts(|| {
-        ERRWRITER.lock().write_fmt(args).unwrap();
-    });
-}
-
-#[doc(hidden)]
-pub fn _print_warn(args: fmt::Arguments) {
-    use core::fmt::Write;
-    use x86_64::instructions::interrupts;
-    interrupts::without_interrupts(|| {
-        WARNWRITER.lock().write_fmt(args).unwrap();
-    });
-}
-
-#[doc(hidden)]
-pub fn _print_info(args: fmt::Arguments) {
-    use core::fmt::Write;
-    use x86_64::instructions::interrupts;
-    interrupts::without_interrupts(|| {
-        INFOWRITER.lock().write_fmt(args).unwrap();
-    });
-}
-
-#[test_case]
-fn test_print_simple() {
-    print!("test_print_simple output\n");
-}
-#[test_case]
-fn test_print_many() {
-    for _ in 0..200 {
-        print!("test_print_many output\n");
-    }
-}
-#[test_case]
-fn test_print_output() {
-    use core::fmt::Write;
-    use x86_64::instructions::interrupts;
-    let s = "Some test string that fits on a single line";
-    interrupts::without_interrupts(|| {
-        let mut writer = WRITER.lock();
-        writeln!(writer, "\n{}", s).expect("writeln failed");
-        for (i, c) in s.chars().enumerate() {
-            let screen_char = writer.buffer.chars[BUFFER_HEIGHT - 2][i].read();
-            assert_eq!(char::from(screen_char.ascii_character), c);
-        }
-    });
-}
