--- rinux/tests/should_panic.rs
@@ -1,25 +1,0 @@
-#![no_std]
-#![no_main]
-
-use rinuxcore::{exit_qemu, serial_print, serial_println, QemuExitCode};
-use core::panic::PanicInfo;
-
-#[no_mangle]
-pub extern "C" fn _start() -> ! {
-    should_fail();
-    serial_println!("[test did not panic]");
-    exit_qemu(QemuExitCode::Failed);
-    loop {}
-}
-
-fn should_fail() {
-    serial_print!("should_panic::should_fail...\t");
-    assert_eq!(0, 1);
-}
-
-#[panic_handler]
-fn panic(_info: &PanicInfo) -> ! {
-    serial_println!("[ok]");
-    exit_qemu(QemuExitCode::Success);
-    loop {}
-}
